refactor: move packages into internal directory
Relocate awsutil, sqsnotify, sqsnotify2 and the stage package under internal/ to mark them as non-importable, and update import paths across main, config, and cmd/sqs-notify2.
