+++ internal/awsutil/aws_windows.go
