+++ internal/sqsnotify2/cache_test.go
@@ -7,7 +7,7 @@ import (
 	"testing"
 
 	"github.com/alicebob/miniredis/v2"
-	"github.com/koron/sqs-notify/sqsnotify2/stage"
+	"github.com/koron/sqs-notify/internal/stage"
 )
 
 func testCache(t *testing.T, c Cache) {
