+++ internal/sqsnotify2/version.go
